Fix Raise Custom Exceptions
In file: traceback_utils.py, method: error_handler, an {Exception} or a {BaseException} was raised. This creates confusion since the exceptions are then differentiated only by the messages they carry. iCR suggested that a specific built-in exception or a custom exception is raised. For more information, read the Python guideline for <hyperlink1>.